Gjort dokumentsidan väldigt mobile-first, det finns fortfarande några saker som ska fixas men det mesta ser bra ut
