--- app/src/Components/Document/ViewDocuments/DocumentCard/DocumentCard.module.css
@@ -1,29 +1,0 @@
-.docCards {
-    padding: 10px;
-    border: 2px solid transparent;
-    border-radius: 5px;
-    text-align: left;
-    cursor:pointer;
-    box-shadow: 0 1px 2px rgba(0,0,0,0.07), 
-                0 2px 4px rgba(0,0,0,0.07), 
-                0 4px 8px rgba(0,0,0,0.07), 
-                0 8px 16px rgba(0,0,0,0.07),
-                0 16px 32px rgba(0,0,0,0.07), 
-                0 32px 64px rgba(0,0,0,0.07);
-    transition: all .31s ease-in-out;
-    width: 230px;
-    margin: 25px auto;
-}
-
-.docCards h3 {
-    margin-left: -1px;
-}
-
-.docCards p {
-    color: #a0a0a0;
-}
-
-.docCards:hover {
-    border: 2px solid #f0c900;
-    transform: scale(1.05);
-}
