+++ app/src/Components/Document/ViewDocuments/DocumentCards/DocumentCards.module.css
@@ -1,9 +1,7 @@
 .DocumentCards {
     display: flex;
     flex-wrap: wrap;
-    width: 80%;
-    margin-left: auto;
-    margin-right: auto;
-    text-align: center;
-    margin-bottom: 20px;
+    width: 97%;
+    justify-content: space-evenly;
+    margin: 10px auto;
 }
