+++ app/src/Components/Document/ViewDocuments/DocumentList/DocumentList.module.css
@@ -1,9 +1,114 @@
-.DocumentList {
-    display: flex;
-    flex-wrap: wrap;
-    width: 80%;
-    margin-left: auto;
-    margin-right: auto;
-    text-align: center;
-    margin-bottom: 20px;
+.DocumentList h3 {
+    text-align: left;
+    margin-left: 10px;
+}
+
+.DocumentList table {
+    margin: 10px auto;
+    text-align: left;
+    width: 97%;
+    border-collapse: separate; 
+    border-spacing: 0 0px;
+    
+}
+
+.DocumentList thead th:hover {
+    color: #ccc;
+}
+
+.DocumentList tr {
+    line-height: 35px;
+}
+
+.DocumentList tr td{
+    border-bottom: 3px solid transparent;
+    border-top: 3px solid transparent;
+}
+
+.DocumentList tbody tr:nth-child(odd) {
+    background: #c4c4c434;
+}
+
+.DocumentList tbody tr:hover td{
+    border-bottom: 3px solid #f0c900;
+    cursor: pointer;
+}
+
+.DocumentList th img {
+    width: 12px;
+}
+
+.arrowUpSelected {
+    border: solid #f0c900;
+    border-width: 0 4px 4px 0;
+    display: inline-block;
+    padding: 3px;
+    transform: rotate(-135deg);
+    -webkit-transform: rotate(-135deg);
+    margin-left: 5px;
+}
+
+.arrowDownSelected {
+    border: solid #f0c900;
+    border-width: 0 4px 4px 0;
+    display: inline-block;
+    padding: 3px;
+    transform: rotate(45deg);
+    -webkit-transform: rotate(45deg);
+    margin-left: 5px;
+}
+
+.arrowUp {
+    border: solid #ccc;
+    border-width: 0 4px 4px 0;
+    display: inline-block;
+    padding: 3px;
+    transform: rotate(-135deg);
+    -webkit-transform: rotate(-135deg);
+    margin-left: 5px;
+}
+
+.arrowDown {
+    border: solid #ccc;
+    border-width: 0 4px 4px 0;
+    display: inline-block;
+    padding: 3px;
+    transform: rotate(45deg);
+    -webkit-transform: rotate(45deg);
+    margin-left: 5px;
+}
+
+.downloadButtonCircle {
+    width: 25px;
+    height: 25px;
+    border: 2px solid #1f1f1f;
+    border-radius: 50%;
+    position: relative;
+    transition: .3s;
+}
+
+.downloadButtonCircle:hover {
+    border-color: #fff;
+}
+
+.downloadButtonCircle:hover .download-btn-arrow{
+    border-color: #fff;
+}
+
+.downloadButtonArrow {
+    border: solid #1f1f1f;
+    position: absolute;
+    left: 7px;
+    top: 5px;
+    border-width: 0 3px 3px 0;
+    display: inline-block;
+    padding: 4px;
+    transform: rotate(45deg);
+    -webkit-transform: rotate(45deg);
+    transition: .3s;
+}
+
+.catParam {
+    cursor: pointer;
+    user-select: none;
 }
